Clean: funcs in cmd_expander, parser and cmd_tab* files
